requirements.txt
langgraph
langgraph-checkpoint-sqlite
langchain-openai
langchain-core
python-dotenv
tavily-python
httpx
beautifulsoup4

